Updates Workflow interface
