--- src/WorkflowInterface.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Ddeboer\DataImport;
-
-/**
- * A mediator between a reader and one or more writers and converters
- *
- * @author Markus Bachmann <markus.bachmann@bachi.biz>
- */
-interface WorkflowInterface
-{
-    /**
-     * Process the whole import workflow
-     *
-     * @return Result Object Containing Workflow Results
-     *
-     * @throws ExceptionInterface
-     */
-    function process();
-}
